% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_unemp}
\alias{clean_unemp}
\title{Cleans unemployment data for Luxembourg}
\usage{
clean_unemp(unemp_data, level, col_of_interest)
}
\arguments{
\item{unemp_data}{A data frame containing unemployment data for Luxembourg.}

\item{level}{A predicate condition indicating the regional level of interest. See details for more.}

\item{col_of_interest}{A column of the `unemp_data` data frame that you wish to select.}
}
\value{
A data frame
}
\description{
Cleans unemployment data for Luxembourg
}
\details{
This function allows the user to create a data frame for several regional levels. The first level
is the complete country. The second level are cantons, and the third level are neither cantons
nor the whole country, so the communes. Individual communes can be selected as well.
`level` must be predicate condition passed down to dplyr::filter. See the examples below
for its usage.
}
\examples{
\dontrun{
# Filter on cantons
clean_unemp(unemp,
            grepl("Canton", commune),
            active_population)
# Filter on a specific commune
clean_unemp(unemp,
            grepl("Kayl", commune),
            active_population)
}
}
